+++ NEWS.md
@@ -1,4 +1,7 @@
 # RecurrenceAnalysis.jl News
+## v1.9.0
+- Specifying metrics with strings in `RecurrenceMatrix` and co is deprecated and will be removed in the next version. Use proper `Metric` instances like `Euclidean()`.
+- Giving a number as `scale` in `RecurrenceMatrix` is deprecated (because its pointless, just give `ε = ε*scale` instead).
 
 ## v1.8.1
 - Better support for iteration over `AbstractRecurrenceMatrix` with functions from `Base`.